move from impls converting from ranges to {Integer,Float}Range
The recently changed builder API takes ranges that convert into these
types, but currently the only conversions available are from themselves.
